\name{vectoptim-package}
\alias{vectoptim-package}
\alias{vectoptim}
\docType{package}
\title{
Vectorized optimization methods
}
\description{
Some vectorized optimization methods and speedup using foreach
}
\details{
\tabular{ll}{
Package: \tab vectoptim\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2014-01-10\cr
License: \tab GPL-3\cr
}

}
\author{
Qiang Kou, Yann Richet
Maintainer: Qiang Kou <qkou@umail.iu.edu>
}


\keyword{ package }
